requirements.txt
numpy
cv2
torchvision
PIL
os
